requirements.txt
fastapi[standard]
pytest
pylint
google-genai
